Moved server DB files to own directory
